edit placeholder, change view on smaller devices, edit error icon, extracted components to different folders
