+++ src/views/studio/MyUploadsView/AssetLine/index.ts
@@ -1,0 +1,1 @@
+export { default as AssetLine } from './AssetLine'
