+++ src/views/studio/MyUploadsView/AssetsGroupUploadBar/AssetsGroupUploadBar.style.tsx
@@ -1,10 +1,5 @@
 import styled from '@emotion/styled'
-import { Text } from '@/shared/components'
 import { colors, sizes, zIndex, breakpoints, transitions } from '@/shared/theme'
-
-type FileLineProps = {
-  isLast?: boolean
-}
 
 type ProgressbarProps = {
   progress: number
@@ -65,8 +60,8 @@ export const Thumbnail = styled.div`
   height: ${sizes(12)};
   background-color: ${colors.gray[700]};
   svg {
-    width: 20px;
-    height: 20px;
+    width: 24px;
+    height: 24px;
   }
   @media screen and (min-width: ${breakpoints.small}) {
     display: flex;
@@ -111,103 +106,11 @@ export const AssetsDrawerContainer = styled.div<DrawerProps>`
   transform: translateY(${({ isActive }) => (isActive ? 0 : '-100%')});
   background-color: ${colors.gray[800]};
   overflow: hidden;
-  transition: all ${transitions.timings.regular} ${transitions.easing};
+  transition: all ${transitions.timings.sharp} ${transitions.easing};
   @media screen and (min-width: ${breakpoints.small}) {
     padding: ${sizes(6)} ${sizes(4)};
   }
   @media screen and (min-width: ${breakpoints.medium}) {
     padding: ${sizes(2)} ${sizes(6)} ${sizes(2)} ${sizes(13)};
-  }
-`
-
-export const FileLineContainer = styled.div<FileLineProps>`
-  display: flex;
-  align-items: center;
-  height: ${sizes(12)};
-  margin-bottom: ${({ isLast }) => !isLast && sizes(12)};
-`
-
-export const StatusMessage = styled(Text)`
-  width: 100%;
-  max-width: ${sizes(16)};
-  height: 48px;
-  margin-left: auto;
-  color: ${colors.gray[300]};
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  @media screen and (min-width: ${breakpoints.medium}) {
-    max-width: ${sizes(42)};
-  }
-`
-
-export const FileLinePoint = styled.div`
-  display: none;
-  width: 37px;
-  height: 32px;
-  border-left: 2px solid ${colors.gray[500]};
-  flex-shrink: 0;
-  &:after {
-    content: '';
-    position: absolute;
-    border-top: 2px solid ${colors.gray[500]};
-    width: 33px;
-    height: 32px;
-    transform: translateY(calc(50% - 1px));
-  }
-  @media screen and (min-width: ${breakpoints.small}) {
-    display: block;
-  }
-`
-
-export const FileLineLastPoint = styled.div`
-  display: none;
-  width: 35px;
-  height: 17px;
-  flex-shrink: 0;
-  border-left: 2px solid ${colors.gray[500]};
-  border-bottom: 2px solid ${colors.gray[500]};
-  transform: translateY(calc(-50% + 1px));
-  @media screen and (min-width: ${breakpoints.small}) {
-    display: block;
-  }
-`
-
-export const FileStatusContainer = styled.div`
-  width: ${sizes(6)};
-  height: ${sizes(6)};
-  margin-left: ${sizes(4)};
-  flex-shrink: 0;
-`
-
-export const FileInfoContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 3fr;
-  width: 100%;
-  max-width: 600px;
-  margin-left: ${sizes(2)};
-  color: ${colors.gray[300]};
-
-  p {
-    display: flex;
-    align-items: center;
-  }
-  @media screen and (min-width: ${breakpoints.small}) {
-    margin-left: ${sizes(6)};
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-template-rows: 1fr;
-  }
-`
-
-export const FileInfoType = styled.div`
-  display: flex;
-  align-items: center;
-  height: 24px;
-  color: ${colors.white};
-  svg {
-    width: 16px;
-    height: 16px;
-    margin-right: 10px;
   }
 `
