+++ src/views/studio/MyUploadsView/AssetsGroupUploadBar/index.ts
@@ -1,0 +1,2 @@
+export { default as AssetsGroupUploadBar } from './AssetsGroupUploadBar'
+export type { UploadData } from './AssetsGroupUploadBar'
